backend/requirements.txt
Flask
pandas
google-generativeai
langchain
python-dotenv
python-docx
youtube-transcript-api
SpeechRecognition
PyPDF2